v9.7.1
* Fix all formatting for golangci-lint
* Move to golangci-lint github action
