+++ exp/func.go
@@ -16,21 +16,32 @@ func (sfe sqlFunctionExpression) Clone() Expression {
 	return sqlFunctionExpression{name: sfe.name, args: sfe.args}
 }
 
-func (sfe sqlFunctionExpression) Expression() Expression                    { return sfe }
-func (sfe sqlFunctionExpression) Args() []interface{}                       { return sfe.args }
-func (sfe sqlFunctionExpression) Name() string                              { return sfe.name }
-func (sfe sqlFunctionExpression) As(val interface{}) AliasedExpression      { return aliased(sfe, val) }
-func (sfe sqlFunctionExpression) Eq(val interface{}) BooleanExpression      { return eq(sfe, val) }
-func (sfe sqlFunctionExpression) Neq(val interface{}) BooleanExpression     { return neq(sfe, val) }
-func (sfe sqlFunctionExpression) Gt(val interface{}) BooleanExpression      { return gt(sfe, val) }
-func (sfe sqlFunctionExpression) Gte(val interface{}) BooleanExpression     { return gte(sfe, val) }
-func (sfe sqlFunctionExpression) Lt(val interface{}) BooleanExpression      { return lt(sfe, val) }
-func (sfe sqlFunctionExpression) Lte(val interface{}) BooleanExpression     { return lte(sfe, val) }
-func (sfe sqlFunctionExpression) Between(val RangeVal) RangeExpression      { return between(sfe, val) }
-func (sfe sqlFunctionExpression) NotBetween(val RangeVal) RangeExpression   { return notBetween(sfe, val) }
+func (sfe sqlFunctionExpression) Expression() Expression { return sfe }
+
+func (sfe sqlFunctionExpression) Args() []interface{} { return sfe.args }
+
+func (sfe sqlFunctionExpression) Name() string { return sfe.name }
+
+func (sfe sqlFunctionExpression) As(val interface{}) AliasedExpression { return aliased(sfe, val) }
+
+func (sfe sqlFunctionExpression) Eq(val interface{}) BooleanExpression  { return eq(sfe, val) }
+func (sfe sqlFunctionExpression) Neq(val interface{}) BooleanExpression { return neq(sfe, val) }
+
+func (sfe sqlFunctionExpression) Gt(val interface{}) BooleanExpression  { return gt(sfe, val) }
+func (sfe sqlFunctionExpression) Gte(val interface{}) BooleanExpression { return gte(sfe, val) }
+func (sfe sqlFunctionExpression) Lt(val interface{}) BooleanExpression  { return lt(sfe, val) }
+func (sfe sqlFunctionExpression) Lte(val interface{}) BooleanExpression { return lte(sfe, val) }
+
+func (sfe sqlFunctionExpression) Between(val RangeVal) RangeExpression { return between(sfe, val) }
+
+func (sfe sqlFunctionExpression) NotBetween(val RangeVal) RangeExpression {
+	return notBetween(sfe, val)
+}
+
 func (sfe sqlFunctionExpression) Like(val interface{}) BooleanExpression    { return like(sfe, val) }
 func (sfe sqlFunctionExpression) NotLike(val interface{}) BooleanExpression { return notLike(sfe, val) }
 func (sfe sqlFunctionExpression) ILike(val interface{}) BooleanExpression   { return iLike(sfe, val) }
+
 func (sfe sqlFunctionExpression) NotILike(val interface{}) BooleanExpression {
 	return notILike(sfe, val)
 }
@@ -46,6 +57,7 @@ func (sfe sqlFunctionExpression) RegexpILike(val interface{}) BooleanExpression 
 func (sfe sqlFunctionExpression) RegexpNotILike(val interface{}) BooleanExpression {
 	return regexpNotILike(sfe, val)
 }
+
 func (sfe sqlFunctionExpression) In(vals ...interface{}) BooleanExpression { return in(sfe, vals...) }
 func (sfe sqlFunctionExpression) NotIn(vals ...interface{}) BooleanExpression {
 	return notIn(sfe, vals...)
