+++ select_dataset_example_test.go
@@ -1,4 +1,4 @@
-// nolint:lll
+// nolint:lll // sql statements are long
 package goqu_test
 
 import (
@@ -118,7 +118,7 @@ func ExampleSelectDataset() {
 
 	sql, args, _ = ds.Prepared(true).ToSQL()
 	fmt.Println(sql, args)
-	// nolint:lll
+	// nolint:lll // SQL statements are long
 	// Output:
 	// SELECT COUNT(*) FROM "test" INNER JOIN "test2" ON ("test"."fkey" = "test2"."id") LEFT JOIN "test3" ON ("test2"."fkey" = "test3"."id") WHERE ((("test"."name" ~ '^(a|b)') AND ("test2"."amount" IS NOT NULL)) AND (("test3"."id" IS NULL) OR ("test3"."status" IN ('passed', 'active', 'registered')))) GROUP BY "test"."user_id" HAVING (AVG("test3"."age") > 10) ORDER BY "test"."created" DESC NULLS LAST []
 	// SELECT COUNT(*) FROM "test" INNER JOIN "test2" ON ("test"."fkey" = "test2"."id") LEFT JOIN "test3" ON ("test2"."fkey" = "test3"."id") WHERE ((("test"."name" ~ ?) AND ("test2"."amount" IS NOT NULL)) AND (("test3"."id" IS NULL) OR ("test3"."status" IN (?, ?, ?)))) GROUP BY "test"."user_id" HAVING (AVG("test3"."age") > ?) ORDER BY "test"."created" DESC NULLS LAST [^(a|b) passed active registered 10]
@@ -615,7 +615,6 @@ func ExampleSelectDataset_Join() {
 	// SELECT * FROM "test" INNER JOIN "test2" USING ("common_column")
 	// SELECT * FROM "test" INNER JOIN (SELECT * FROM "test2" WHERE ("amount" > 0)) ON ("test"."fkey" = "test2"."Id")
 	// SELECT * FROM "test" INNER JOIN (SELECT * FROM "test2" WHERE ("amount" > 0)) AS "t" ON ("test"."fkey" = "t"."Id")
-
 }
 
 func ExampleSelectDataset_InnerJoin() {
@@ -1228,7 +1227,7 @@ func ExampleSelectDataset_Prepared() {
 		"col5": []string{"a", "b", "c"},
 	}).ToSQL()
 	fmt.Println(sql, args)
-	// nolint:lll
+	// nolint:lll // sql statements are long
 	// Output:
 	// SELECT * FROM "items" WHERE (("col1" = ?) AND ("col2" = ?) AND ("col3" IS TRUE) AND ("col4" IS FALSE) AND ("col5" IN (?, ?, ?))) [a 1 a b c]
 }
@@ -1496,7 +1495,6 @@ func ExampleSelectDataset_ScanVals() {
 }
 
 func ExampleSelectDataset_ScanVal() {
-
 	db := getDb()
 	findUserIDByName := func(name string) {
 		var id int64
@@ -1523,7 +1521,6 @@ func ExampleSelectDataset_ScanVal() {
 }
 
 func ExampleSelectDataset_Count() {
-
 	if count, err := getDb().From("goqu_user").Count(); err != nil {
 		fmt.Println(err.Error())
 	} else {
